Moved embeds to their own files; added props validation
