+++ react/gameday2/gameday2.js
@@ -4,6 +4,9 @@ var classNames = require('classnames');
 var _ = require('underscore');
 
 import GamedayNavbar from './components/GamedayNavbar';
+import EmbedTwitch from './components/EmbedTwitch';
+import Embedyoutube from './components/EmbedYoutube';
+import EmbedUstream from './components/EmbedUstream';
 
 var GamedayFrame = React.createClass({
   getInitialState: function() {
@@ -325,67 +328,6 @@ var VideoCellOverlay = React.createClass({
         </div>
       )
     }
-  }
-});
-
-var EmbedYoutube = React.createClass({
-  render: function() {
-    var src = "//www.youtube.com/embed/" + this.props.webcast.channel;
-    return (
-      <iframe
-        width={this.props.vidWidth}
-        height={this.props.vidHeight}
-        src={src}
-        frameBorder="0"
-        allowFullScreen>
-      </iframe>
-    );
-  }
-});
-
-var EmbedUstream = React.createClass({
-  render: function() {
-    var channel = this.props.webcast.channel;
-    var src = `http://www.ustream.tv/embed/${channel}?html5ui=1`;
-    return (
-      <iframe
-        width={this.props.vidWidth}
-        height={this.props.vidHeight}
-        src={src}
-        scrolling="no"
-        allowfullscreen
-        webkitallowfullscreen
-        frameborder="0"
-        style={{border: "0 none transparent"}}></iframe>
-    );
-  }
-});
-
-var EmbedTwitch = React.createClass({
-  render: function() {
-    var channel = this.props.webcast.channel;
-    var src = `http://www.twitch.tv/widgets/live_embed_player.swf?channel=${channel}`;
-    var flashvars = `hostname=www.twitch.tv&channel=${channel}&auto_play=true&start_volume=25&enable_javascript=true`;
-    return (
-      <object
-        type='application/x-shockwave-flash'
-        height='100%'
-        width='100%'
-        id='live_embed_player_flash'
-        data={src}
-        bgcolor='#000000'>
-        <param name='allowFullScreen' value='true' />
-        <param name='allowScriptAccess' value='always' />
-        <param name='allowNetworking' value='all' />
-        <param
-          name='movie'
-          value='http://www.twitch.tv/widgets/live_embed_player.swf' />
-        <param
-          name='flashvars'
-          value={flashvars} />
-        <param name='wmode' value='transparent' />
-      </object>
-    );
   }
 });
